requirements.txt
uvicorn
fastapi
python-multipart
python-dotenv
openai
ultralytics
transformers